requirements.txt
gunicorn
uvicorn
fastapi
openai
pypdf
python-dotenv
python-multipart


